// TPT/manage-tradingpt/app/api/users.ts
export interface LevelTestInfo {
  status: string | null;
  grade: string | null;
  gradingTrainerName: string | null;
}

export interface NewSubscriptionCustomer {
  customerId: number;
  name: string;
  phoneNumber: string;
  uid: string | null;
  hasAttemptedLevelTest: boolean;
  levelTestInfo: LevelTestInfo | null;
  hasConsultation: boolean;
  assignedTrainerName: string | null;
}